+++ CodinGame/CodeRoyale/CodeRoyale.rs
@@ -112,6 +112,8 @@ fn closest_free_site(sites: &HashMap<i32, Site>, coord: (i32, i32)) -> i32 {
     min_id
 }
 
+// TODO: unit-training function returning a string based on gold
+
 fn main() {
     let mut input_line = String::new();
     io::stdin().read_line(&mut input_line).unwrap();
@@ -130,6 +132,8 @@ fn main() {
 
         sites.insert(site_id, Site::new(x, y, radius));
     }
+
+    let mut barracks = "".to_string();
 
     // game loop
     loop {
@@ -151,6 +155,10 @@ fn main() {
             let param_1 = parse_input!(inputs[5], i32);
             let param_2 = parse_input!(inputs[6], i32);
 
+            if structure_type == 2 {
+                barracks.push_str(&format!("{} ", site_id));
+            }
+
             sites.get_mut(&site_id).unwrap().
                 update(structure_type, owner, param_1, param_2);
         }
@@ -160,7 +168,6 @@ fn main() {
         let num_units = parse_input!(input_line, i32);
 
         let mut units = vec![];
-
         for i in 0..num_units as usize {
             let mut input_line = String::new();
             io::stdin().read_line(&mut input_line).unwrap();
@@ -192,11 +199,17 @@ fn main() {
 
         let queen_loc = queen_location(&units);
         let closest = closest_free_site(&sites, queen_loc);
+        let cl_xy = sites.get(&closest).unwrap().get_location();
 
         eprintln!("Queen on: {:?}", queen_loc);
-        eprintln!("Closest site: {:?}", closest);
+        eprintln!("Closest site: {:?} on {},{}", closest, cl_xy.0, cl_xy.1);
 
-        println!("WAIT");
-        println!("TRAIN");
+        if touched_site == closest {
+            println!("BUILD {} BARRACKS-KNIGHT", closest);
+        } else {
+            println!("MOVE {} {}", cl_xy.0, cl_xy.1);
+        }
+
+        println!("TRAIN {}", barracks);
     }
 }